[BUG] 김진도 - NPC 메인페이지 버튼 오류 fix
