+++ src/main/resources/templates/npc/main.html
@@ -12,16 +12,15 @@
         .btn-group { display: flex; flex-direction: column; gap: 15px; margin-top: 30px; }
         .btn { padding: 20px; font-size: 18px; font-weight: bold; border-radius: 8px; text-decoration: none; display: block; border: 2px solid #000; transition: all 0.3s; }
 
-        /* 문의하기 버튼: 기본적으로 검정색 바탕 */
         .btn-inquire { background: #000; color: #fff; }
         .btn-inquire:hover { background: #333; }
 
-        /* 신고하기 버튼: 흰색 바탕에 검정 글씨 */
         .btn-report { background: #fff; color: #000; }
         .btn-report:hover { background: #f0f0f0; }
 
         .alert-msg { background: #e8f5e9; color: #2e7d32; padding: 15px; border-radius: 5px; margin-bottom: 20px; font-weight: bold; }
     </style>
+    <th:block th:replace="~{modal/modal :: modalCss}"></th:block>
 </head>
 <body>
 <div class="container">
@@ -33,7 +32,6 @@
 
     <div class="btn-group">
         <a href="/npc/inquiry" class="btn btn-inquire">1:1 문의하기 ➔</a>
-
         <a href="/npc/my-inquiries" class="btn" style="background:#f4f4f4; color:#333; border: 1px solid #ccc;">내 문의 및 답변 확인</a>
 
         <a href="/npc/report" class="btn btn-report">⚠️ 불량 사용자 신고하기</a>
@@ -41,8 +39,11 @@
     </div>
 
     <div style="margin-top: 30px;">
-        <a href="//" style="color: #999; text-decoration: underline;">메인 화면으로 돌아가기</a>
+        <a th:href="${role == 'TEACHER' ? '/teacher/main' : '/student/main'}" style="color: #999; text-decoration: underline;">메인 화면으로 돌아가기</a>
     </div>
 </div>
+
+<div th:replace="~{modal/modal :: modalHtml}"></div>
+<script th:replace="~{modal/modal :: modalScript}"></script>
 </body>
 </html>
